requirements.txt
pydantic>=2.7          # v2 family 
openai>=1.80
aiolimiter>=1.2
 
langchain-core>=0.3.1          # first LangChain fully on Pydantic v2
langchain-community>=0.3.1
llama-index-core>=0.12.0    
 
neo4j>=5.20
graspologic>=3.4
networkx>=3.3       
 
llama-index-graph-stores-neo4j>=0.4.0       
llama-index-graph-stores-nebula>=0.4.0      
 
llama-index-llms-openai>=0.4.0            
llama-index-llms-azure-openai>=0.3.0      
llama-index-llms-databricks>=0.3.0        
llama-index-embeddings-huggingface>=0.5.0   
llama-index-embeddings-openai>=0.3.0       
llama-index-embeddings-azure-openai>=0.3.0  
